% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ryx}
\alias{plot.ryx}
\title{plot.ryx}
\usage{
\method{plot}{ryx}(x, ...)
}
\arguments{
\item{x}{The output of ryx()}

\item{...}{Any additional passed arguments}
}
\value{
A plot showing the absolute value of the correlation coefficients
}
\description{
Generic plot function specified for the ryx() function
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 plot(x)
 }
}
}
