% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ryx}
\alias{print.ryx}
\title{print.ryx}
\usage{
\method{print}{ryx}(x, ...)
}
\arguments{
\item{x}{The output of ryx()}

\item{...}{Any additional passed arguments}
}
\value{
A table showing the output of the ryx() function
}
\description{
Generic print function specified for the ryx() function
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 print(x)
 }
}
}
