% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{ryx}
\alias{ryx}
\title{ryx}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{Dataframe containing variables}

\item{y}{Variable to measure correlations against}

\item{x}{Variable(s) to test against y}
}
\value{
A list of length 3. 1. The y variables. 2. The x variable(s). 3. A dataframe containing the correlation measures with significance checks.
}
\description{
Create correlation measure between two or more variables
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 library(MASS)
 x <- ryx(Boston, y = "medv")
 }
}
}
