% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryxhw4-package.R
\docType{package}
\name{ryxhw4-package}
\alias{ryxhw4}
\alias{ryxhw4-package}
\title{ryxhw4: What the Package Does (One Line, Title Case)}
\description{
This package is designed to measure the correlation between a specified variable and any number of additional variables. It consists of the core ryx() function to generate correlations as well as generic print, summary, and plot functions to display the output.
}
\author{
\strong{Maintainer}: Johnny Wilkens \email{jwilkens@wesleyan.edu}

}
\keyword{internal}
