% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ryx}
\alias{summary.ryx}
\title{summary.ryx}
\usage{
\method{summary}{ryx}(x, ...)
}
\arguments{
\item{x}{The output of ryx()}

\item{...}{Any additional passed arguments}
}
\value{
A summary of the output of the ryx() function in words.
}
\description{
Generic summary function specified for the ryx() function
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 summary(x)
 }
}
}
